release: v2.0.0 – April 4, 2025

+++ server/src/inquiry/inquiry.controller.ts
@@ -8,13 +8,14 @@ import {
   Delete,
   UseGuards,
   Request,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { InquiryService } from './inquiry.service';
 import { CreateInquiryDto } from './dto/create-inquiry.dto';
 import { UpdateInquiryDto } from './dto/update-inquiry.dto';
 import { ApiBody, ApiOperation, ApiTags } from '@nestjs/swagger';
 import { Inquiry } from './entities/inquiry.entity';
-import { JwtAuthGuard } from 'src/auth/jwt-auth/jwt-auth.guard';
+import { AccessTokenGuard } from 'src/auth/guards/access-token.guard';
 
 @ApiTags('Inquiries')
 @Controller('inquiries')
@@ -22,14 +23,18 @@ export class InquiryController {
   constructor(private readonly inquiryService: InquiryService) {}
 
   @Post()
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '문의사항 작성(JWT 인증 필요)',
+    summary: '문의사항 작성(Admin JWT 인증 필요)',
     description: '문의사항을 작성합니다.',
   })
   @ApiBody({ type: CreateInquiryDto })
   create(@Request() req, @Body() createInquiryDto: CreateInquiryDto) {
-    return this.inquiryService.create(createInquiryDto, req);
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to create inquiries.');
+    }
+
+    return this.inquiryService.create(req.user.id, createInquiryDto);
   }
 
   @Get()
@@ -51,25 +56,32 @@ export class InquiryController {
   }
 
   @Patch(':inquiryId')
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '해당 ID 문의사항 업데이트(JWT 인증 필요)',
+    summary: '해당 ID 문의사항 업데이트(Admin JWT 인증 필요)',
     description: '해당 ID의 문의사항을 업데이트합니다.',
   })
   update(
+    @Request() req,
     @Param('inquiryId') inquiryId: string,
-    @Body() updateInquiryDto: UpdateInquiryDto
+    @Body() updateInquiryDto: UpdateInquiryDto,
   ) {
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to update inquiries.');
+    }
     return this.inquiryService.update(inquiryId, updateInquiryDto);
   }
 
   @Delete(':inquiryId')
-  @UseGuards(JwtAuthGuard)
+  @UseGuards(AccessTokenGuard)
   @ApiOperation({
-    summary: '해당 ID 문의사항 삭제(JWT 인증 필요)',
+    summary: '해당 ID 문의사항 삭제(Admin JWT 인증 필요)',
     description: '해당 ID의 문의사항을 삭제합니다.',
   })
-  remove(@Param('inquiryId') inquiryId: string) {
+  remove(@Request() req, @Param('inquiryId') inquiryId: string) {
+    if (!req.user.isAdmin) {
+      throw new UnauthorizedException('Only Admins are allowed to delete inquiries.');
+    }
     return this.inquiryService.remove(inquiryId);
   }
 }
